pyproject.toml
[project]
name = "aipocalypse"
version = "0.1.0"
description = "Agentic playground for simulating the zombie apocalypse"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "devtools>=0.12.2",
    "dill>=0.4.0",
    "dotenv>=0.9.9",
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "neomodel>=5.5.3",
    "toonify>=1.4.0",
]
